ocr_mcp/pyproject.toml
[project]
name = "ocr-mcp"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "mcp[cli]>=1.0,<2",
    "pydantic>=2.6,<3",
    "security",
]

[dependency-groups]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5,<6",
    "mypy>=1.9",
    "ruff>=0.4",
    "Pillow>=10",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ocr_mcp"]

[tool.uv.sources]
security = { path = "../security", editable = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--tb=short --cov=ocr_mcp --cov-report=term-missing"

[tool.mypy]
python_version = "3.12"
packages = ["ocr_mcp"]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S", "C901", "SIM", "RET", "N"]
ignore = ["S101"]
